Only expose render fn in modmod, add cli
