+++ modmod/Cargo.toml
@@ -6,8 +6,12 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+clap = { version = "4.3.11", features = ["derive"], optional = true }
 fs_extra = "1.3.0"
 globset = { version = "0.4.11", default-features = false }
 indoc = "2.0.2"
 serde = { version = "1.0.166", features = ["derive"] }
 toml = { version = "0.7.6", default-features = false, features = ["parse"] }
+
+[features]
+default = ["dep:clap"]
