+++ modmod/src/error.rs
@@ -1,11 +1,10 @@
 use std::io;
 
-
 #[derive(Debug)]
 pub enum Error {
     Io(io::Error),
     Toml(toml::de::Error),
-    Output(OutputError)
+    Output(OutputError),
 }
 
 impl From<io::Error> for Error {
