+++ modmod/src/load.rs
@@ -6,7 +6,10 @@ use crate::{Module, PathTo, Topic, Track};
 
 pub trait Load: DeserializeOwned + Sized {
     fn load(path: &Path, base_path: Option<&Path>) -> crate::Result<PathTo<Self>> {
-        let path = base_path.map(|b| b.join(path)).unwrap_or(path.to_owned()).canonicalize()?;
+        let path = base_path
+            .map(|b| b.join(path))
+            .unwrap_or(path.to_owned())
+            .canonicalize()?;
         let content = fs::read_to_string(&path)?;
         let data = toml::from_str(&content)?;
